DataTablesColumnDirective can be used as child component for DataTablesComponent to extract column options. DataTablesComponent accepts data and rendering options. DemoComponent creates now datatable in static and dynamic mode.
