+++ src/app/pages/demo/demo.component.spec.ts
@@ -78,9 +78,9 @@ describe('Demo Component', () => {
     // WHEN clicking third tab
     clickTab(2);
 
-    // THEN datatables plugin should be visible
+    // THEN datatables plugin(s) should be visible
     fixture.whenStable().then(() => {
-        expect(fixture.nativeElement.querySelectorAll('dng2-dt').length).toEqual(1);
+        expect(fixture.nativeElement.querySelectorAll('dng2-dt').length).toEqual(2);
     });
   }));
