+++ src/app/widgets/datatables/datatables.component.html
@@ -1,8 +1,6 @@
-<h1>DataTables</h1>
-    <table #dtElemDynamic class="table table-striped table-bordered" cellspacing="0"></table>
-    <br/>
-    <table #dtElemStatic class="table table-striped table-bordered" cellspacing="0">
-        <tbody>
-            <tr *ngFor="let row of staticData;"><td *ngFor="let key of staticDataOrder;">{{ createColumnHTML(row, key) }}</td></tr>
-        </tbody>
-    </table>
+<table *ngIf="!staticRender" #dtElem class="table table-striped table-bordered" cellspacing="0"></table>
+<table *ngIf="staticRender" #dtElem class="table table-striped table-bordered" cellspacing="0">
+  <tbody>
+    <tr *ngFor="let row of tableData;"><td *ngFor="let key of tableOptions.order;">{{ createColumnHTML(row, key) }}</td></tr>
+  </tbody>
+</table>
