+++ src/app/widgets/datatables/datatables.component.ts
@@ -1,4 +1,4 @@
-import { Component, ViewChild, ElementRef, AfterViewInit } from '@angular/core';
+import { Component, ViewChild, ElementRef, AfterViewInit, Directive, ContentChildren, QueryList, Input } from '@angular/core';
 import 'jquery';
 import 'datatables.net';
 declare var $: any;
@@ -15,100 +15,60 @@ declare var $: any;
 // import 'datatables.net-bs';
 import './datatables.bootstrap4';
 
-const tableDataOrder = ['name', 'title', 'location', 'salary'];
-const tableData = [
-    { name: 'Tiger Nixon', title: 'System Architect', location: 'Edinburgh', salary: '$320,800' },
-    { name: 'Garrett Winters', title: 'Accountant', location: 'Tokyo', salary: '$170,750' },
-    { name: 'Ashton Cox', title: 'Junior Technical Author', location: 'San Francisco', salary: '$86,000' },
-    { name: 'Cedric Kelly', title: 'Senior Javascript Developer', location: 'Edinburgh', salary: '$433,060' },
-    { name: 'Airi Satou', title: 'Accountant', location: 'Tokyo', salary: '$162,700' },
-    { name: 'Brielle Williamson', title: 'Integration Specialist', location: 'New York', salary: '$372,000' },
-    { name: 'Herrod Chandler', title: 'Sales Assistant', location: 'San Francisco', salary: '$137,500' },
-    { name: 'Rhona Davidson', title: 'Integration Specialist', location: 'Tokyo', salary: '$327,900' },
-    { name: 'Colleen Hurst', title: 'Javascript Developer', location: 'San Francisco', salary: '$205,500' },
-    { name: 'Sonya Frost', title: 'Software Engineer', location: 'Edinburgh', salary: '$103,600' },
-    { name: 'Jena Gaines', title: 'Office Manager', location: 'London', salary: '$90,560' },
-    { name: 'Quinn Flynn', title: 'Support Lead', location: 'Edinburgh', salary: '$342,000' },
-    { name: 'Charde Marshall', title: 'Regional Director', location: 'San Francisco', salary: '$470,600' },
-    { name: 'Haley Kennedy', title: 'Senior Marketing Designer', location: 'London', salary: '$313,500' },
-    { name: 'Tatyana Fitzpatrick', title: 'Regional Director', location: 'London', salary: '$385,750' },
-    { name: 'Michael Silva', title: 'Marketing Designer', location: 'London', salary: '$198,500' },
-    { name: 'Paul Byrd', title: 'Chief Financial Officer (CFO)', location: 'New York', salary: '$725,000' },
-    { name: 'Gloria Little', title: 'Systems Administrator', location: 'New York', salary: '$237,500' },
-    { name: 'Bradley Greer', title: 'Software Engineer', location: 'London', salary: '$132,000' },
-    { name: 'Dai Rios', title: 'Personnel Lead', location: 'Edinburgh', salary: '$217,500' },
-    { name: 'Jenette Caldwell', title: 'Development Lead', location: 'New York', salary: '$345,000' },
-    { name: 'Yuri Berry', title: 'Chief Marketing Officer (CMO)', location: 'New York', salary: '$675,000' },
-    { name: 'Caesar Vance', title: 'Pre-Sales Support', location: 'New York', salary: '$106,450' },
-    { name: 'Doris Wilder', title: 'Sales Assistant', location: 'Sidney', salary: '$85,600' },
-    { name: 'Angelica Ramos', title: 'Chief Executive Officer (CEO)', location: 'London', salary: '$1,200,000' },
-    { name: 'Gavin Joyce', title: 'Developer', location: 'Edinburgh', salary: '$92,575' },
-    { name: 'Jennifer Chang', title: 'Regional Director', location: 'Singapore', salary: '$357,650' },
-    { name: 'Brenden Wagner', title: 'Software Engineer', location: 'San Francisco', salary: '$206,850' },
-    { name: 'Fiona Green', title: 'Chief Operating Officer (COO)', location: 'San Francisco', salary: '$850,000' },
-    { name: 'Shou Itou', title: 'Regional Marketing', location: 'Tokyo', salary: '$163,000' },
-    { name: 'Michelle House', title: 'Integration Specialist', location: 'Sidney', salary: '$95,400' },
-    { name: 'Suki Burks', title: 'Developer', location: 'London', salary: '$114,500' },
-    { name: 'Prescott Bartlett', title: 'Technical Author', location: 'London', salary: '$145,000' },
-    { name: 'Gavin Cortez', title: 'Team Leader', location: 'San Francisco', salary: '$235,500' },
-    { name: 'Martena Mccray', title: 'Post-Sales support', location: 'Edinburgh', salary: '$324,050' },
-    { name: 'Unity Butler', title: 'Marketing Designer', location: 'San Francisco', salary: '$85,675' }
-];
+
+/**
+ * Simple directive that is used to indicate column data for the datatable component.
+ */
+@Directive({
+    selector: '[dng2-dt-column-directive]',
+})
+export class DataTablesColumnDirective {
+
+    constructor(private el: ElementRef) {}
+
+    getAttributes(): any {
+        const attr = {};
+
+        const attributes = this.el.nativeElement.attributes;
+        for (let i = 0; i < attributes.length; i++) {
+            if (attributes[i].name.startsWith('dt-')) {
+                attr[attributes[i].name.replace('dt-', '')] = attributes[i].value;
+            }
+        }
+
+        return attr;
+    }
+}
 
 @Component({
     selector: 'dng2-dt',
     template: require('./datatables.component.html')
 })
 export class DataTablesComponent implements AfterViewInit {
-    @ViewChild('dtElemDynamic') elDynamic: ElementRef;
-    @ViewChild('dtElemStatic') elStatic: ElementRef;
+    @Input() tableData?: any;
+    @Input() tableOptions?: any;
+    @Input() staticRender? = false;
+    @ViewChild('dtElem') el: ElementRef;
+    @ContentChildren(DataTablesColumnDirective) rows: QueryList<DataTablesColumnDirective>;
 
-    protected staticData: Array<any> = tableData;
-    protected staticDataOrder: Array<string> = tableDataOrder;
+    protected options = {};
 
     ngAfterViewInit() {
 
-        // Create DataTables from static HTML (Angular first creates table HTML that includes all rows and columns)
-        $(this.elStatic.nativeElement).DataTable({
-            columns: [
-                {
-                    title: 'Name',
-                },
-                {
-                    title: 'Position',
-                },
-                {
-                    title: 'Office',
-                },
-                {
-                    title: 'Salary',
-                }
-            ]
+        // Get the datatable column attributes from child components
+        const rowAttr = this.rows.map(row => {
+            return row.getAttributes();
         });
 
-        // Create DataTables from dynamic HTML (DataTables created all HTML inside <table> tag)
-        const options: any = {
-            columns: [
-                {
-                    title: 'Name',
-                    data: 'name'
-                },
-                {
-                    title: 'Position',
-                    data: 'title'
-                },
-                {
-                    title: 'Location',
-                    data: 'location'
-                },
-                {
-                    title: 'Salary',
-                    data: 'name'
-                }
-            ],
-            data: tableData
-        };
-        $(this.elDynamic.nativeElement).DataTable(options);
+        this.options['columns'] = rowAttr;
+
+        // Unless the table HTML is created, DataTables needs the data for the table
+        if (!this.staticRender) {
+            this.options['data'] = this.tableData;
+        }
+
+        // Create DataTables
+        $(this.el.nativeElement).DataTable(this.options);
     }
 
     protected createColumnHTML(data: any, key: string): string {
